requirements.txt
anthropic>=0.40.0
PyGithub>=2.1.1
gitpython>=3.1.40
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
requests>=2.31.0
playwright>=1.42.0
